Fix C++ container type to use long when numeric_literal_suffix=AUTO

+++ tests/integration/cases/int_list/Cpp_numeric_literal_suffix_auto.cpp
@@ -5,7 +5,7 @@ struct _Any {
     _Any(std::initializer_list<_Any>) noexcept {}
 };
 void _check() {
-_Any my_data = std::vector<int>{
+_Any my_data = std::vector<long>{
     1L,
     2L,
     3L,
